Added menu functions, uploaded dictionary files
Updating PassGen-Dev as follows:

Refined the New-Char function to use a single array of characters and use the toLower() method as needed to change case. The upper-case character pool was used to avoid using characters which are visually confusing - the letters i, l and o will not be used.

Added 3 function declarations - a native .NET GUI, a ShowUI GUI, and a text-based menu.

Uploaded the dictionary text files used by this script.

Script can be invoked by dot-loading and running the commands "Native-GUI", "ShowUI-GUI" or "Text-Menu".